refactor: using mapTokenHistory from utils
